Develop

+++ client/src/services/user/deleteUser.js
@@ -1,15 +1,21 @@
 const deleteUser = (userInfos) => {
-  const userId = localStorage.getItem('userId')
+  const userId = window.location.href.split('/update/')[1]
+  const token = localStorage.getItem('token')
   fetch('http://localhost:8000/api/user/' + userId, {
     method: 'delete',
     headers: {
       'Content-Type': 'application/json',
+      authorization: 'Bearer ' + token,
     },
     body: JSON.stringify(userInfos),
   })
-    .then(() => {
-      localStorage.clear()
-      window.location.reload()
+    .then((res) => {
+      if (res.status === 200) {
+        localStorage.clear()
+        window.location.reload()
+      } else {
+        alert('Email / mot de passe incorrect')
+      }
     })
     .catch((error) => console.log(error))
 }
